Updated tests to run in current project structure
